% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(path = ".", id = NULL, file = "codemeta.json")
}
\arguments{
\item{path}{path to the package root}

\item{id}{identifier for the package (e.g. a DOI, as URL)}

\item{file}{output file location, should be called \code{codemeta.json}. Set to NULL to supress writing file.}
}
\value{
a codemeta list object (invisbly) and write out the codemeta.json file
}
\description{
write_codemeta
}
\examples{
\dontrun{
write_codemeta()
}
}
