% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codemeta.R
\name{write_codemeta}
\alias{write_codemeta}
\title{write_codemeta}
\usage{
write_codemeta(
  path = ".",
  id = NULL,
  file = "codemeta.json",
  verbose = getOption("verbose", FALSE)
)
}
\arguments{
\item{path}{path to the package root}

\item{id}{identifier for the package (e.g. a DOI, as URL)}

\item{file}{output file location, should be called \code{codemeta.json}.}

\item{verbose}{Whether to print messages indicating the progress of internet
downloads.}
}
\value{
a codemeta list object (invisbly) and write out the codemeta.json file
}
\description{
write_codemeta
}
\examples{
\donttest{

 # 'path' and 'out' here are for illustrative use only.
 # typical use in a package is simply `wite_codemeta()` with no arguments

 path <- system.file("", package="codemeta")
 out <- tempfile(fileext =".json")
 write_codemeta(path, file = out)
}
}
